+++ src/gopro/meta.rs
@@ -46,27 +46,27 @@ impl GoProMeta {
         Ok(meta)
     }
 
-    fn muid() -> Result<Vec<u32>, GpmfError> {
-        let fourcc = FourCC::from_str("MUID");
+    // fn muid() -> Result<Vec<u32>, GpmfError> {
+    //     let fourcc = FourCC::from_str("MUID");
 
-        // for field in self.udta.iter() {
-        //     if field.name == fourcc {
-        //         let no_of_entries = match ((field.size - 8) % 4, (field.size - 8) / 4) {
-        //             (0, n) => n,
-        //             (_, n) => panic!("Failed to determine MUID: {n} length field is not 32-bit aligned")
-        //         };
+    //     // for field in self.udta.iter() {
+    //     //     if field.name == fourcc {
+    //     //         let no_of_entries = match ((field.size - 8) % 4, (field.size - 8) / 4) {
+    //     //             (0, n) => n,
+    //     //             (_, n) => panic!("Failed to determine MUID: {n} length field is not 32-bit aligned")
+    //     //         };
 
-        //         let mut fld = field.to_owned();
+    //     //         let mut fld = field.to_owned();
 
-        //         return (0..no_of_entries).into_iter()
-        //                 .map(|_| fld.data.read_le::<u32>()) // read LE to match GPMF
-        //                 .collect::<BinResult<Vec<u32>>>()
-        //                 .map_err(|err| GpmfError::BinReadError(err))
-        //     }
-        // }
+    //     //         return (0..no_of_entries).into_iter()
+    //     //                 .map(|_| fld.data.read_le::<u32>()) // read LE to match GPMF
+    //     //                 .collect::<BinResult<Vec<u32>>>()
+    //     //                 .map_err(|err| GpmfError::BinReadError(err))
+    //     //     }
+    //     // }
 
-        Ok(Vec::new())
-    }
+    //     Ok(Vec::new())
+    // }
 
-    fn gumi() {}
+    // fn gumi() {}
 }