+++ src/lib.rs
@@ -44,6 +44,7 @@ mod content_types;
 mod gopro;
 mod constants;
 mod types;
+mod tests;
 
 pub use gpmf::{
     Gpmf,